# docs/curl/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartLogistics",
    "description": "\nBackend for **SmartLogistics** \u2014 TransFleet's supply chain and delivery\norchestration platform.\n\nAuthenticate at `POST /api/v1/auth/login`, then use **Authorize** to send the\naccess token with subsequent requests.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate and receive a token pair",
        "description": "Repeated failures lock the account temporarily.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bad credentials, locked or inactive account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Exchange a refresh token for a new pair",
        "description": "Refresh tokens rotate: the presented token is revoked and replaced.\n\nReplaying an already-rotated token revokes every session for that user, on\nthe assumption it was stolen.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Token invalid, expired or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke the current session, or all sessions",
        "description": "Omit ``refresh_token`` to sign out of every device.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "List the caller's active sessions",
        "operationId": "list_sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionRead"
                  },
                  "type": "array",
                  "title": "Response List Sessions Api V1 Auth Sessions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Return the authenticated user's profile",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/inventory": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "List stock levels",
        "description": "`available_qty` is a generated column \u2014 on hand minus held \u2014 so it can\nnever disagree with its inputs.",
        "operationId": "list_stock_api_v1_inventory_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          },
          {
            "name": "sku_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sku Id"
            }
          },
          {
            "name": "below_reorder_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only positions at or under their reorder threshold.",
              "title": "Below Reorder Level"
            },
            "description": "Only positions at or under their reorder threshold."
          },
          {
            "name": "in_stock_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "In Stock Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InventoryItemRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/reservations": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "List stock holds",
        "description": "Shows what stock is held, for whom, and when the hold lapses.",
        "operationId": "list_reservations_api_v1_inventory_reservations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shipment Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReservationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationRead"
                  },
                  "title": "Response List Reservations Api V1 Inventory Reservations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/release-expired": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Release stock holds that have lapsed",
        "description": "Expired holds on contended stock are released automatically when someone\ntries to reserve it. This sweeps the rest, and is what Celery beat will call\non a schedule in Phase 3 \u2014 correctness does not depend on it running.",
        "operationId": "release_expired_api_v1_inventory_release_expired_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Release Expired Api V1 Inventory Release Expired Post"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted for your role or warehouse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "example": {
          "email": "admin@transfleet.com",
          "password": "SmartLogistics!2026"
        }
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access token lifetime in seconds."
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "LoginResponse",
        "description": "Tokens plus the profile, so the client avoids an immediate second call."
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token",
            "description": "Omit to revoke every session for the current user."
          }
        },
        "type": "object",
        "title": "LogoutRequest"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "Page_InventoryItemRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InventoryItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total matching records, ignoring pagination."
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[InventoryItemRead]"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable, machine-readable error identifier."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation."
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Field-level errors, when applicable."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Problem",
        "description": "RFC 7807-style error body used for every non-2xx response."
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "ReservationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "inventory_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Inventory Item Id"
          },
          "shipment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Shipment Id"
          },
          "warehouse_id": {
            "type": "string",
            "format": "uuid",
            "title": "Warehouse Id"
          },
          "sku_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sku Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "committed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "release_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "inventory_item_id",
          "shipment_id",
          "warehouse_id",
          "sku_id",
          "quantity",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "ReservationRead"
      },
      "ReservationStatus": {
        "type": "string",
        "enum": [
          "held",
          "committed",
          "released"
        ],
        "title": "ReservationStatus",
        "description": "Lifecycle of an inventory hold."
      },
...
      "SessionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "expires_at"
        ],
        "title": "SessionRead",
        "description": "An active refresh-token session, for the 'where am I logged in' view."
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access token lifetime in seconds."
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPair",
        "description": "Issued on login and on refresh. Refresh tokens rotate on every use."
      },
...
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Login identifier. Stored lowercased."
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "pattern": "^\\+?[0-9\\s\\-()]{7,32}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "E.164 preferred. Used for courier and support contact."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "id",
          "role",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "UserRead",
        "description": "Full representation returned to admins and to the user themselves."
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "customer_support",
          "warehouse_operator",
          "courier"
        ],
        "title": "UserRole",
        "description": "Roles defined by the SmartLogistics access model."
      },
...
      "UserStatus": {
        "type": "string",
        "enum": [
          "pending_activation",
          "active",
          "suspended",
          "deactivated"
        ],
        "title": "UserStatus",
        "description": "Account lifecycle state, independent of authentication outcome."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token",
        "scheme": "bearer"
      }
    }
  },
...
}